// packages/i-npm-commander/src/types/NpmPackage.ts
type NpmSemverString = string
type FormattedSemverString = string
type NpmUTC0DateString = string

interface NpmPackageDistTarballInfo {
    /**
     * @description one sha512 value to validate the package's integrity
     */
    integrity: string
    shasum: string
    tarball: string
    /**
     * @description fileCount in tarball
     */
    fileCount: number
    /**
     * @description size after tarball unpacked
     */
    unpackedSize: number/*  | bigint */
    /**
     * @description signature for this version tarball generated by npm
     */
    'npm-signature': string
}

export interface NpmPackageInfoAsDependency {
    /**
     * @type semver
     */
    version: string
    /**
     * @description npm package's verification id, generally generated by sha512 algorithm
     * 
     * @sample "sha512-a9gxpmdXtZEInkCSHUJDLHZVBgb1QS0jhss4cPP93EW7s+uC5bikET2twEF3KV+7rDblJcmNvTR7VJejqd2C2g=="
     */
    integrity: string,
    /**
     * @description whether used as devDependencies in npm project
     */
    dev: true,
    requires?: {
        [pkgname: string]: string/* semver */
    }
}

/**
 * @description structure to describte the FULL information about one package,
 * in general, returned by `https://registry.npmjs.com/[PKG]` WITHOUT header [referer]
 */
type NpmPackageInfoForExplorer = Pick<PackageJSON, 
    'name'
    | 'version'
    | 'description'
    | 'main'
    | 'bin'
    | 'scripts'
    | 'repository'
    | 'keywords'
    | 'author'
    | 'license'
    | 'bugs'
    | 'homepage'
    | 'dependencies'
    | 'devDependencies'
    | 'engines'
> & {
    dist: NpmPackageDistTarballInfo
    types?: string
    gitHead?: string
    maintainers?: NpmUserInfo[]
    directories?: Record<string, string>
    readme?: string
    readmeFilename?: string
    [any_other_user_defined_key: string]: any
}

export type NpmPackageInfoFromBrowser = Omit<NpmPackageIndexedCriticalInfo, 'versions'> & NpmPackageInfoForExplorer & {
    versions: {
        [formatted_semver_key: string]: NpmPackageInfoForExplorer & {
            /* npm specified fields, pointless in general :start */
            _id?: string
            _nodeVersion?: string
            _npmVersion?: string
            _npmUser?: NpmUserInfo
            _npmOperationalInternal?: {
                host: string
                tmp: string
            }
            _hasShrinkwrap?: boolean
            /* npm specified fields, pointless in general :end */
            [any_other_user_defined_key: string]: any
        }
    }
    time: {
        created: NpmUTC0DateString
        modified: NpmUTC0DateString
        [formatted_semver_key: string]: NpmUTC0DateString
    }
} 

/**
 * @description structure to describte the CRITICAL information about one package,
 * in general, returned by `https://registry.npmjs.com/[PKG]` with header [referer: 'install [PKG]']
 */
export interface NpmPackageIndexedCriticalInfo {
	name: string,
	"dist-tags": {
        /**
         * @sample "latest": "0.8.4"
         */
        [tagname: string]: FormattedSemverString
	},
	versions: {
        [formatted_semver_key: string]: {
            name: NpmPackageIndexedCriticalInfo['name']
            version: FormattedSemverString
            dist: NpmPackageDistTarballInfo
            dependencies?: {
                /**
                 * @sample "@fibjs/mkdirp": "^1.0.1",
                 */
                [pkgname: string]: NpmSemverString
            },
            devDependencies?: NpmPackageIndexedCriticalInfo['versions'][any]['dependencies'],
            bin?: {
                /**
                 * @sample "fib-typify": "./bin/fib-typify.js"
                 */
                [bin_name: string]: string
            },
            engines?: {
                /**
                 * @sample "fibjs": ">=0.28.0"
                 */
                [engine_name: string]: string
            },
        }
    },
    /**
     * @description latest modified date string
     */
	modified: NpmUTC0DateString
}